Move the selectors to the global directory
