+++ client/post-editor/editor-title/index.jsx
@@ -17,7 +17,7 @@ import TextareaAutosize from 'components/textarea-autosize';
 import { isMobile } from 'lib/viewport';
 import * as stats from 'lib/posts/stats';
 import { getSelectedSiteId } from 'state/ui/selectors';
-import { areSitePermalinksEditable } from 'state/sites/selectors';
+import { areSitePermalinksEditable } from 'state/selectors';
 import { isEditorNewPost, getEditorPostId } from 'state/ui/editor/selectors';
 import { getEditedPost } from 'state/posts/selectors';
 import { editPost } from 'state/posts/actions';
