+++ client/state/posts/selectors.js
@@ -13,7 +13,6 @@ import {
 	getSerializedPostsQuery,
 	getDeserializedPostsQueryDetails,
 	getSerializedPostsQueryWithoutPage,
-	isEmptyContent,
 	mergeIgnoringArrays,
 	normalizePostForEditing,
 	normalizePostForDisplay
@@ -361,31 +360,6 @@ export const isEditedPostDirty = createSelector(
 	},
 	( state ) => [ state.posts.items, state.posts.edits ]
 );
-
-/**
-  * Returns true if the edited post has content
-  * (title, excerpt or content not empty)
-  *
-  * @param  {Object}  state  Global state tree
-  * @param  {Number}  siteId Site ID
-  * @param  {Number}  postId Post ID
-  * @return {Boolean}        Whether the edited post has content or not
- */
-export function editedPostHasContent( state, siteId, postId ) {
-	const editedPost = getEditedPost( state, siteId, postId );
-	return (
-		!! editedPost &&
-		(
-			some( [ 'title', 'excerpt' ], ( field ) => editedPost[ field ] && !! editedPost[ field ].trim() ) ||
-			/*
-			 * We don't yet have the notion of post's raw content in the Redux state so we rely on post content attribute here
-			 * when we do, we'll want it to reflect the Flux implementation's emptiness check
-			 * where raw content is preferred to the content property if available
-			 */
-			! isEmptyContent( editedPost.content )
-		)
-	);
-}
 
 /**
  * Returns true if the post status is publish, private, or future
